+++ BUILD
@@ -1,23 +1,23 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 load("@bazel_gazelle//:def.bzl", "gazelle")
 
-# gazelle:prefix github.com/discentem/pantri_but_go
+# gazelle:prefix github.com/discentem/cavorite
 # gazelle:exclude .sl
 gazelle(name = "gazelle")
 
 go_library(
-    name = "pantri_but_go_lib",
+    name = "cavorite_lib",
     srcs = ["main.go"],
-    importpath = "github.com/discentem/pantri_but_go",
+    importpath = "github.com/discentem/cavorite",
     visibility = ["//visibility:private"],
     deps = [
-        "//internal/cmd/pantri",
+        "//internal/cmd/root",
         "@com_github_google_logger//:logger",
     ],
 )
 
 go_binary(
-    name = "pantri_but_go",
-    embed = [":pantri_but_go_lib"],
+    name = "cavorite",
+    embed = [":cavorite_lib"],
     visibility = ["//visibility:public"],
 )
