+++ internal/cmd/root/root.go
@@ -1,4 +1,4 @@
-package pantri
+package root
 
 import (
 	"errors"
