+++ internal/metadata/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "metadata",
     srcs = ["metadata.go"],
-    importpath = "github.com/discentem/pantri_but_go/internal/metadata",
+    importpath = "github.com/discentem/cavorite/internal/metadata",
     visibility = ["//:__subpackages__"],
     deps = ["@com_github_spf13_afero//:afero"],
 )
