+++ internal/stores/s3.go
@@ -18,7 +18,7 @@ import (
 	"github.com/google/logger"
 	"github.com/spf13/afero"
 
-	"github.com/discentem/pantri_but_go/internal/metadata"
+	"github.com/discentem/cavorite/internal/metadata"
 
 	s3manager "github.com/aws/aws-sdk-go-v2/feature/s3/manager"
 )
